backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.9.2
pydantic[email]==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2

# AI / RAG
langgraph==0.2.34
langchain==0.3.3
langchain-anthropic==0.2.3
langchain-huggingface==0.1.0
langchain-groq==0.2.1
anthropic==0.37.1
openai==1.51.0
tavily-python==0.3.9
chromadb==0.5.11
sentence-transformers==3.1.1
pypdf==5.0.1
ragas==0.2.3
datasets==3.0.1
